Renamed model_input to modelinput.
