+++ neural_networks/chessbot/modelinput.py
